trucksimulator-telemetry-python/requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
ply==3.11
robotpy-cppheaderparser==5.0.16
wincertstore==0.2
